portable-ruby: fix more compiler toolchain references
We already get rid of the compiler shim references in `rbconfig.rb`, but
there are still some more that may not be valid on the target system:

    $ brew ruby -e 'p RbConfig::CONFIG.select { |k, v| v.match? /gcc-(.*)-\d+/ }'
    {"RANLIB"=>"gcc-ranlib-11", "NM"=>"gcc-nm-11", "AR"=>"gcc-ar-11"}

We can replace them with `ranlib`, `nm`, and `ar`, respectively.

This should help to avoid issues like the one spotted in
Homebrew/brew#17114.